Fixes environment closing and tests for task implementations (#231)
# Description

This MR fixes the test scripts inside the `omni.isaac.orbit_tasks`
extension. The tests were previously crashing since they weren't updated
to latest changes in the core framework.

Additionally, the tests helped in realizing that the environment did not
close properly, i.e. some managers do not get destroyed properly and
still hold references to certain scene entities. The MR now explicitly
calls all `del` operators over all the managers to delete them in the
right-order.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

---------

+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/terrains/terrain_generator.py
@@ -150,7 +150,7 @@ class TerrainGenerator:
         for index in range(self.cfg.num_cols):
             sub_index = np.min(np.where(index / self.cfg.num_cols + 0.001 < np.cumsum(proportions))[0])
             sub_indices.append(sub_index)
-        sub_indices = np.array(sub_indices, dtype=np.int)
+        sub_indices = np.array(sub_indices, dtype=np.int32)
         # create a list of all terrain configs
         sub_terrains_cfgs = list(self.cfg.sub_terrains.values())
